+++ vector/src/main/java/im/vector/riotx/core/preference/UserAvatarPreference.kt
@@ -23,6 +23,8 @@ import android.widget.ProgressBar
 import androidx.preference.Preference
 import androidx.preference.PreferenceViewHolder
 import im.vector.matrix.android.api.session.Session
+import im.vector.matrix.android.api.util.MatrixItem
+import im.vector.matrix.android.api.util.toMatrixItem
 import im.vector.riotx.R
 import im.vector.riotx.core.extensions.vectorComponent
 import im.vector.riotx.features.home.AvatarRenderer
@@ -59,9 +61,9 @@ open class UserAvatarPreference : Preference {
         val session = mSession ?: return
         val view = mAvatarView ?: return
         session.getUser(session.myUserId)?.let {
-            avatarRenderer.render(it, view)
+            avatarRenderer.render(it.toMatrixItem(), view)
         } ?: run {
-            avatarRenderer.render(null, session.myUserId, null, view)
+            avatarRenderer.render(MatrixItem.UserItem(session.myUserId), view)
         }
     }