+++ vector/src/main/java/im/vector/riotx/features/home/HomeDetailFragment.kt
@@ -27,6 +27,7 @@ import com.google.android.material.bottomnavigation.BottomNavigationItemView
 import com.google.android.material.bottomnavigation.BottomNavigationMenuView
 import im.vector.matrix.android.api.session.crypto.keysbackup.KeysBackupState
 import im.vector.matrix.android.api.session.group.model.GroupSummary
+import im.vector.matrix.android.api.util.toMatrixItem
 import im.vector.riotx.R
 import im.vector.riotx.core.extensions.commitTransactionNow
 import im.vector.riotx.core.platform.ToolbarConfigurable
@@ -74,12 +75,7 @@ class HomeDetailFragment @Inject constructor(
 
     private fun onGroupChange(groupSummary: GroupSummary?) {
         groupSummary?.let {
-            avatarRenderer.render(
-                    it.avatarUrl,
-                    it.groupId,
-                    it.displayName,
-                    groupToolbarAvatarImageView
-            )
+            avatarRenderer.render(it.toMatrixItem(), groupToolbarAvatarImageView)
         }
     }
 
@@ -155,7 +151,7 @@ class HomeDetailFragment @Inject constructor(
         bottomNavigationView.selectedItemId = when (displayMode) {
             RoomListDisplayMode.PEOPLE -> R.id.bottom_action_people
             RoomListDisplayMode.ROOMS  -> R.id.bottom_action_rooms
-            else                                -> R.id.bottom_action_home
+            else                       -> R.id.bottom_action_home
         }
     }