+++ vector/src/main/java/im/vector/riotx/features/home/room/breadcrumbs/BreadcrumbsFragment.kt
@@ -67,4 +67,8 @@ class BreadcrumbsFragment @Inject constructor(
     override fun onBreadcrumbClicked(roomId: String) {
         sharedActionViewModel.post(RoomDetailSharedAction.SwitchToRoom(roomId))
     }
+
+    fun scrollToTop() {
+        breadcrumbsRecyclerView.scrollToPosition(0)
+    }
 }