+++ vector/src/main/java/im/vector/riotx/features/home/room/detail/RoomDetailFragment.kt
@@ -66,6 +66,8 @@ import im.vector.matrix.android.api.session.room.timeline.Timeline
 import im.vector.matrix.android.api.session.room.timeline.TimelineEvent
 import im.vector.matrix.android.api.session.room.timeline.getLastMessageContent
 import im.vector.matrix.android.api.session.user.model.User
+import im.vector.matrix.android.api.util.MatrixItem
+import im.vector.matrix.android.api.util.toMatrixItem
 import im.vector.riotx.R
 import im.vector.riotx.core.dialogs.withColoredButton
 import im.vector.riotx.core.epoxy.LayoutManagerStateRestorer
@@ -408,9 +410,7 @@ class RoomDetailFragment @Inject constructor(
         composerLayout.sendButton.setContentDescription(getString(descriptionRes))
 
         avatarRenderer.render(
-                event.senderAvatar,
-                event.root.senderId ?: "",
-                event.getDisambiguatedDisplayName(),
+                MatrixItem.UserItem(event.root.senderId ?: "", event.getDisambiguatedDisplayName(), event.senderAvatar),
                 composerLayout.composerRelatedMessageAvatar
         )
         composerLayout.expand {
@@ -599,20 +599,19 @@ class RoomDetailFragment @Inject constructor(
                         }
 
                         // Replace the word by its completion
-                        val displayName = item.displayName ?: item.userId
+                        val matrixItem = item.toMatrixItem()
+                        val displayName = matrixItem.getBestName()
 
                         // with a trailing space
                         editable.replace(startIndex, endIndex, "$displayName ")
 
                         // Add the span
-                        val user = session.getUser(item.userId)
                         val span = PillImageSpan(
                                 glideRequests,
                                 avatarRenderer,
                                 requireContext(),
-                                item.userId,
-                                user?.displayName ?: item.userId,
-                                user?.avatarUrl)
+                                matrixItem
+                        )
                         span.bind(composerLayout.composerEditText)
 
                         editable.setSpan(span, startIndex, startIndex + displayName.length, Spannable.SPAN_EXCLUSIVE_EXCLUSIVE)
@@ -684,7 +683,7 @@ class RoomDetailFragment @Inject constructor(
             inviteView.visibility = View.GONE
             val uid = session.myUserId
             val meMember = session.getRoom(state.roomId)?.getRoomMember(uid)
-            avatarRenderer.render(meMember?.avatarUrl, uid, meMember?.displayName, composerLayout.composerAvatarImageView)
+            avatarRenderer.render(MatrixItem.UserItem(uid, meMember?.displayName, meMember?.avatarUrl), composerLayout.composerAvatarImageView)
         } else if (summary?.membership == Membership.INVITE && inviter != null) {
             inviteView.visibility = View.VISIBLE
             inviteView.render(inviter, VectorInviteView.Mode.LARGE)
@@ -711,7 +710,7 @@ class RoomDetailFragment @Inject constructor(
                 activity?.finish()
             } else {
                 roomToolbarTitleView.text = it.displayName
-                avatarRenderer.render(it, roomToolbarAvatarImageView)
+                avatarRenderer.render(it.toMatrixItem(), roomToolbarAvatarImageView)
                 roomToolbarSubtitleView.setTextOrHide(it.topic)
             }
             jumpToBottomView.count = it.notificationCount
@@ -1195,9 +1194,8 @@ class RoomDetailFragment @Inject constructor(
                                             glideRequests,
                                             avatarRenderer,
                                             requireContext(),
-                                            userId,
-                                            displayName,
-                                            roomMember?.avatarUrl)
+                                            MatrixItem.UserItem(userId, displayName, roomMember?.avatarUrl)
+                                    )
                                             .also { it.bind(composerLayout.composerEditText) },
                                     0,
                                     displayName.length,