+++ vector/src/main/java/im/vector/riotx/features/home/room/detail/readreceipts/DisplayReadReceiptsController.kt
@@ -21,6 +21,7 @@ import im.vector.matrix.android.api.session.Session
 import im.vector.riotx.core.date.VectorDateFormatter
 import im.vector.riotx.features.home.AvatarRenderer
 import im.vector.riotx.features.home.room.detail.timeline.item.ReadReceiptData
+import im.vector.riotx.features.home.room.detail.timeline.item.toMatrixItem
 import javax.inject.Inject
 
 /**
@@ -36,9 +37,7 @@ class DisplayReadReceiptsController @Inject constructor(private val dateFormatte
             val timestamp = dateFormatter.formatRelativeDateTime(it.timestamp)
             DisplayReadReceiptItem_()
                     .id(it.userId)
-                    .userId(it.userId)
-                    .avatarUrl(it.avatarUrl)
-                    .name(it.displayName)
+                    .matrixItem(it.toMatrixItem())
                     .avatarRenderer(avatarRender)
                     .timestamp(timestamp)
                     .addIf(session.myUserId != it.userId, this)